Allow libxml parser options to be passed to `parse`, `parseFile`
The motivation for this change was to correct the behavior of XSD
includes/imports (Issue #11). Paths to included schemas should be resolved
relative to the XSD file, not the execution directory. This was only a
problem for the top level XSD - nested includes were resolved relative to
the files that included them. Setting the `baseUrl` option to the path of
the top level XSD file fixes the unexpected behavior.

Tests have been updated so that they work with includes relative to the
schema files.